+++ wikipedia-db/Rust/preprocessor/src/main.rs
@@ -3,9 +3,9 @@ use std::fs::File;
 use preprocessor::run;
 
 fn main() {
-	let text_cat = File::open("test_samples/nrm/nrmwiki-20210201-page.sql")
+	let text_cat = File::open("../../eo/eowiki-20210201-page.sql")
 		.expect("Something went wrong reading the file one_line_categories");
-	let text_links = File::open("test_samples/nrm/nrmwiki-20210201-categorylinks.sql")
+	let text_links = File::open("../../eo/eowiki-20210201-categorylinks.sql")
 		.expect("Something went wrong reading the file one_line_categorieslinks");
 
 	run(text_cat, text_links)
