+++ wikipedia-db/Rust/preprocessor/tests/eo.rs
@@ -1,12 +1,13 @@
 use std::fs::File;
 
-use crate::run;
+use preprocessor::run;
 
 #[test]
-fn norman() {
-	let text_cat = File::open("../../nrm/nrmwiki-20210201-page.sql")
+#[ignore]
+fn test_eo() {
+	let text_cat = File::open("../../eo/eowiki-20210201-page.sql")
 		.expect("Something went wrong reading the file one_line_categories");
-	let text_links = File::open("../../nrm/nrmwiki-20210201-categorylinks.sql")
+	let text_links = File::open("../../eo/eowiki-20210201-categorylinks.sql")
 		.expect("Something went wrong reading the file one_line_categorieslinks");
 
 	run(text_cat, text_links)
